+++ src/components/WorkOrder/PrintJobTicketDetails.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import WarningText from '../Template/WarningText'
 import { WorkOrder } from '../../models/workOrder'
 import { formatDateTime } from 'src/utils/time'
-import { getCautionaryAlertsType } from '../../utils/cautContactAlerts'
+import { getCautionaryAlertsType } from '../../utils/cautionaryAlerts'
 
 const PrintJobTicketDetails = ({
   workOrder,
